+++ src/System.Windows.Forms/tests/UnitTests/System/Windows/Forms/ScreenTests.cs
@@ -33,15 +33,15 @@ public class ScreenTests
 
     public static IEnumerable<object[]> Equals_Screen_TestData()
     {
-        Screen screen = new((HMONITOR)1);
+        Screen screen = new((HMONITOR)(nint)1);
         yield return new object[] { screen, screen, true };
-        yield return new object[] { screen, new Screen((HMONITOR)1), true };
-        yield return new object[] { screen, new Screen((HMONITOR)2), false };
+        yield return new object[] { screen, new Screen((HMONITOR)(nint)1), true };
+        yield return new object[] { screen, new Screen((HMONITOR)(nint)2), false };
     }
 
     public static IEnumerable<object[]> Equals_Object_TestData()
     {
-        Screen screen = new((HMONITOR)1);
+        Screen screen = new((HMONITOR)(nint)1);
         yield return new object[] { screen, new(), false };
         yield return new object[] { screen, null, false };
     }
